requirements.txt
chompjs>=1.3.0
httpx>=0.28.1
PyYAML>=6.0.2
